CRUD Product.cs finalizado

+++ distributed-system/Entities/Product.cs
@@ -5,6 +5,6 @@ public class Product
     public int Id { get; set; }
     public string Brand { get; set; }
     public string Name { get; set; }
-    public decimal CostUnit { get; set; }
-    public decimal SalePrice { get; set; }
+    public float CostUnit { get; set; }
+    public float SalePrice { get; set; }
 }
